v2/requirements.txt
pandas>=2.0.0
typer>=0.9.0  
rich>=13.0.0
pyyaml>=6.0
mutagen>=1.47.0
nltk>=3.8.0
python-Levenshtein>=0.21.0